Refactor common fetch logic into fetch_utils
